flask_service/requirements.txt
Flask==2.2.3  # Ensure you're using a compatible Flask version
Werkzeug==2.2.3  # Compatible version for Flask 2.x
Flask-Cors>=4.0.0
Flask-Admin>=1.6.1
Flask-PyMongo>=2.3.0
mongoengine>=0.27
python-dotenv>=1.0.0
Pillow>=10.0.0  # For image processing if needed
gunicorn>=21.2.0
flask-mongoengine==1.0.0
